requirements/base.txt
Django==4.1.3
psycopg2==2.9.5
django-environ==0.9.0
django-extensions>=3.2.1
django-cors-headers==3.13.0
Pillow==9.4.0

#Djanfo filter
django-filter==21.1
djangorestframework-filters==1.0.0.dev2

# restframework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0

# documentation
drf-yasg==1.21.4

# Async
uvicorn[standard]==0.19.0

# PDF
pdfkit==1.0.0

# Pandas
pandas==1.5.3
numpy==1.24.0
django-pandas==0.6.6
XlsxWriter==3.0.7

# Celery
celery==5.2.7
flower==1.2.0
redis==4.5.5

# monitorin
prometheus-client==0.19.0
